faltanpocoscambios
cambio en tipogasto

+++ PrototipoMRP/CP_gastoV2.cs
@@ -52,13 +52,22 @@ namespace PrototipoMRP
 
         private void button3_Click(object sender, EventArgs e)
         {
-            MessageBox.Show("LOS DATOS SERAN BORRADOS", "Cuidado", MessageBoxButtons.OKCancel, MessageBoxIcon.Information);
-            tipoGastov1.eliminarTipoGasto(textBox4.Text); 
+            if(
+            MessageBox.Show("LOS DATOS SERAN BORRADOS", "Cuidado", MessageBoxButtons.YesNo)==DialogResult.Yes)
+            {
+                tipoGastov1.eliminarTipoGasto(textBox4.Text);
+                limpiarTipoG("");
+                consulta();
+                MessageBox.Show("LOS DATOS SE BORRARON");
+            }
+
         }
 
         private void button5_Click(object sender, EventArgs e)
         {
-            tipoGastov1.insertarTipoGasto( textBox2.Text, textBox3.Text);
+            tipoGastov1.insertarTipoGasto(textBox4.Text, textBox2.Text, textBox3.Text);
+            limpiarTipoG("");
+            consulta();
         }
 
         private void button2_Click(object sender, EventArgs e)
